refactor: publish pkgs; add FinalizeValidationResult
