+++ pkg/sinks/sink.go
