+++ pkg/sinks/sink_test.go
